Not mine (#3)
* Update README.md

Replaced a leftover dns-netcup with dns-ispconfig in Named arguments section. Also added a ISPConfig 3.1 path to the Credentials file, that is required, I found it to fail when not included.

* Update dns_ispconfig.py

Added the `stamp` parameter to API JSON that is required by ISPConfig

* Allow full name zones

Add the original record name (full domain including _acme-challenge bit) to the list of domain guesses to check with ISPConfig. Now the full record is tried first and so if there is a zone named _acme-challenge.example.com it will be used instead of example.com. The actual record name is then empty.

* Update sutup.py

Removed encoding parameter as it is not known to python 2.7 and default in python 3

Co-authored-by: Michal Krejčí <michal@krejci.vip>